Move Footer into folder
