+++ src/components/Footer/index.js
@@ -1,0 +1,5 @@
+/**
+ * Re-export to prevent nested path
+ */
+import Footer from './Footer'
+export default Footer
